requirements.txt
pyparsing==3.0.9
pyyaml==6.0
mako==1.2.3
pandas==1.4.4
